move PageLayout into layout directory
